Upload mathsat bin + minors
